requirements.txt
asttokens==2.2.1
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
decorator==5.1.1
executing==1.2.0
flake8==6.0.0
greenlet==2.0.2
idna==3.4
ipython==8.14.0
jedi==0.18.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy==1.4.1
mypy-extensions==1.0.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.15.1
python-dotenv==1.0.0
requests==2.31.0
ruff==0.0.275
six==1.16.0
SQLAlchemy==2.0.17
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.7.0
urllib3==2.0.3
vk-api==11.9.9
wcwidth==0.2.6
